Remove CommandServiceIdExtractor
