+++ src/Command/InstanceIsReadyCommand.php
@@ -8,7 +8,6 @@ use App\Exception\ServiceIdMissingException;
 use App\Services\CommandActionRunner;
 use App\Services\CommandConfigurator;
 use App\Services\CommandInstanceRepository;
-use App\Services\CommandServiceIdExtractor;
 use App\Services\InstanceClient;
 use App\Services\ServiceConfiguration;
 use DigitalOceanV2\Exception\ExceptionInterface;
@@ -21,7 +20,7 @@ use Symfony\Component\Console\Output\OutputInterface;
     name: InstanceIsReadyCommand::NAME,
     description: 'Check if an instance is ready to be used',
 )]
-class InstanceIsReadyCommand extends Command
+class InstanceIsReadyCommand extends AbstractServiceCommand
 {
     use RetryableCommandTrait;
 
@@ -35,14 +34,13 @@ class InstanceIsReadyCommand extends Command
     public const DEFAULT_RETRY_DELAY = 30;
 
     public function __construct(
-        private CommandConfigurator $configurator,
-        private CommandServiceIdExtractor $serviceIdExtractor,
+        CommandConfigurator $configurator,
         private InstanceClient $instanceClient,
         private CommandActionRunner $commandActionRunner,
         private CommandInstanceRepository $commandInstanceRepository,
         private ServiceConfiguration $serviceConfiguration,
     ) {
-        parent::__construct();
+        parent::__construct($configurator);
     }
 
     protected function configure(): void
@@ -51,7 +49,6 @@ class InstanceIsReadyCommand extends Command
 
         $this->configurator
             ->addId($this)
-            ->addServiceIdOption($this)
             ->addRetryLimitOption($this, self::DEFAULT_RETRY_LIMIT)
             ->addRetryDelayOption($this, self::DEFAULT_RETRY_DELAY)
         ;
@@ -63,7 +60,7 @@ class InstanceIsReadyCommand extends Command
      */
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $serviceId = $this->serviceIdExtractor->extract($input);
+        $serviceId = $this->getServiceId($input);
         $serviceConfiguration = $this->serviceConfiguration->getServiceConfiguration($serviceId);
         if (null === $serviceConfiguration) {
             $output->write('No configuration for service "' . $serviceId . '"');
