+++ src/Command/ServiceConfigurationExistsCommand.php
@@ -6,7 +6,6 @@ namespace App\Command;
 
 use App\Exception\ServiceIdMissingException;
 use App\Services\CommandConfigurator;
-use App\Services\CommandServiceIdExtractor;
 use App\Services\ServiceConfiguration;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
@@ -17,16 +16,15 @@ use Symfony\Component\Console\Output\OutputInterface;
     name: ServiceConfigurationExistsCommand::NAME,
     description: 'Check if configuration exists for a given service_id',
 )]
-class ServiceConfigurationExistsCommand extends Command
+class ServiceConfigurationExistsCommand extends AbstractServiceCommand
 {
     public const NAME = 'app:service-configuration:exists';
 
     public function __construct(
-        private CommandConfigurator $configurator,
-        private CommandServiceIdExtractor $serviceIdExtractor,
+        CommandConfigurator $configurator,
         private ServiceConfiguration $serviceConfiguration,
     ) {
-        parent::__construct();
+        parent::__construct($configurator);
     }
 
     protected function configure(): void
@@ -41,7 +39,7 @@ class ServiceConfigurationExistsCommand extends Command
      */
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $serviceId = $this->serviceIdExtractor->extract($input);
+        $serviceId = $this->getServiceId($input);
 
         return $this->serviceConfiguration->exists($serviceId) ? Command::SUCCESS : Command::FAILURE;
     }
